Loading images

Loading images

Data types and Interface

Load tiff-file

Computing and applying image mask

Generation of control images

Data types and Interface

All images need to be represented as the mutable struct MultiChannelImage for further use with ProteinCoLoc.jl. This MultiChannelImage allows storing all pixel itensities, channel names, the image name, path to the individual channel images, the size of the image as well as the Otsu's threshold for each color channel.

ProteinCoLoc.MultiChannelImage — Type

mutable struct MultiChannelImage{T <: Union{Missing, Float64}, S <: AbstractString, F <: AbstractFloat}

This mutable struct represents a multi-channel image.

Fields

data::Vector{Matrix{T}}: A Vector of Matrix{T} representing the image data for each channel.

channels::Vector{S}: A Vector of strings representing the names of the channels.

name::S: A string representing the name of the image.

path::Vector{S}: A Vector of strings representing the paths to the image files for each channel.

pixel_size::Tuple{F, F}: A Tuple of two Float64 representing the width and height of the pixels in micrometers.

otsu_threshold::Vector{F}: A Vector of Float64 representing the Otsu threshold for each channel.

Constructor

The constructor checks that the lengths of the data, channels, and otsu_threshold vectors match.

Notes

This struct is used to store and manipulate multi-channel images in the ProteinCoLoc package.

A MultiChannelImage can be created either manually or by calling the external constructor that loads tiff files into a MultiChannelImage

ProteinCoLoc.MultiChannelImage — Method

MultiChannelImage(name::S, path::Vector{S}, channels::Vector{S} =[]) where {S <: AbstractString}

This constructor creates a new MultiChannelImage.

Arguments

name: A string representing the name of the image.

path: A Vector of strings representing the paths to the image files for each channel.

channels: A Vector of strings representing the names of the channels. Default is an empty Vector.

Returns

img: A MultiChannelImage representing the loaded image.

Notes

This constructor loads the image data from the files at the specified paths, checks if channel names are provided and generates default names if not, calculates the pixel size and Otsu threshold for each channel, and creates a new MultiChannelImage with these values.

Several instances of type MultiChannelImage can be stored as a MultiChannelImageStack{T <:MultiChannelImage, S <: AbstractString}. For example images taken from the samples of the same biological condition can be grouped in an MultiChannelImageStackfor further downstream analysis. A MultiChannelImageStackcan be generated by using the constructor MultiChannelImageStack(img, name)

ProteinCoLoc.MultiChannelImageStack — Type

mutable struct MultiChannelImageStack{T <: MultiChannelImage, S <: AbstractString}

This mutable struct represents a stack of multi-channel images.

Fields

img: A Vector of MultiChannelImage representing the images in the stack.

name: A string representing the name of the image stack.

num_images: An Int64 representing the number of images in the stack.

Constructor

The constructor takes a Vector of MultiChannelImage and a string as arguments, and sets the num_images field to the length of the Vector.

Notes

This struct is used to store and manipulate stacks of multi-channel images in the ProteinCoLoc package.

The following methods are defined for an object of type MultiChannelImageStack:

getindex(stack, i): Method to retrieve the MultiChannelImage at index i of the MultiChannelImageStack stack.

ìterate(stack, state): Method to iterate over the images in the MultiChannelImageStack stack. This method definition allows the use of for-loops etc with MultiChannelImageStacks.

Base.getindex — Method

Base.getindex(stack::MultiChannelImageStack, i::Int64)

Function to retrieve an image from a MultiChannelImageStack.

Arguments

stack: A MultiChannelImageStack from which to retrieve an image.

i: An Int64 representing the index of the image to retrieve.

Returns

img: A MultiChannelImage representing the image at index i in the stack.

Errors

Throws a BoundsError if i is greater than the number of images in the stack.

Notes

This function checks that i is within the bounds of the stack, and returns the image at index i.

Base.iterate — Method

Base.iterate(stack::MultiChannelImageStack, state=1)

This function overrides the base iterate function to iterate over the images in a MultiChannelImageStack.

Arguments

stack: A MultiChannelImageStack over which to iterate.

state: An Int64 representing the current state of the iteration. Default is 1.

Returns

A tuple containing the current image and the next state, or nothing if the end of the stack has been reached.

Notes

This function checks if the state is greater than the number of images in the stack, and if so, returns nothing. Otherwise, it returns the current image and the next state.

Load tiff-file

Images can be loaded from a file using the function 'load_tiff'. This function converts the RGB image into a greyscale image and returns a Matrix{Float64}

ProteinCoLoc.load_tiff — Method

load_tiff(path::S) where {S<:AbstractString}

This function loads a TIFF image from a file and converts it to a matrix of Float64 values, where the values represent the intensity in grayscale.

Arguments

path: A string representing the path to the TIFF image file.

Returns

img: A Matrix of Float64 representing the grayscale intensity of the image.

Notes

This function uses the Images.jl package to load the image, convert it to grayscale, and then convert it to a matrix.

Computing and applying image mask

The image mask using Otsu thresholding can be computed and directly applied to an img of type MultiChannelImage using the function apply_mask!(img).

ProteinCoLoc.apply_mask! — Function

apply_mask!(img)

This function calculates a  mask for a MultiChannelImage using Otsu thresholding and directly applies it to the image.

# Arguments
- `img`: A MultiChannelImage to which to apply the mask.

# Returns
- `img`: The MultiChannelImage with the mask applied.

_calculate_mask and _apply_mask! are not part of the public API

In addition, two low-level function are available to perform these steps individually. These functions are not part of the public API and might change in the future.

ProteinCoLoc._calculate_mask — Function

_calculate_mask(img::MultiChannelImage)

This function calculates a binary mask for each channel of a MultiChannelImage, where pixels with intensity greater than the Otsu threshold are set to 1 and others are set to 0.

Arguments

img: A MultiChannelImage for which to calculate the mask.

Returns

mask: A Vector of Matrix{Bool} representing the binary mask for each channel.

Notes

This function iterates over the channels of the image, applies the Otsu threshold to the data of each channel, and stores the resulting binary mask.

ProteinCoLoc._apply_mask! — Method

_apply_mask!(img::MultiChannelImage, mask::Vector{Matrix{Bool}})

This function applies a binary mask to each channel of a MultiChannelImage, where pixels in the mask that are set to 0 are set to 0 in the image.

Arguments

img: A MultiChannelImage to which to apply the mask.

mask: A Vector of Matrix{Bool} representing the binary mask for each channel.

Returns

img: The MultiChannelImage with the mask applied.

Notes

This function iterates over the channels of the image, multiplies the data of each channel by the corresponding mask, and stores the result back in the image.

apply_mask!(img::MultiChannelImage)

This function calculates a binary mask for a MultiChannelImage and applies it to the image.

Arguments

img: A MultiChannelImage to which to apply the mask.

Returns

img: The MultiChannelImage with the mask applied.

Notes

This function calls the _calculate_mask function to calculate the mask, and then calls the _apply_mask! function to apply the mask to the image.

Generation of control images

ProteinCoLocs provides two different ways to generate artificial control images. As discussed in our paper, the use of randomly shuffled or block shuffled image are not ideal as these may lead to unreliable results. Therefore, we strongly recommend the use of dedicated control images (e.g. control staining without primary antibodies) that have been taken with the same settings.

The first image randomly shuffles all pixels of each color channel.

ProteinCoLoc.shuffle_pixels — Function

shuffle_pixels(img::MultiChannelImage)

This function shuffles the pixel values in each channel of a MultiChannelImage.

Arguments

img: A MultiChannelImage whose pixel values are to be shuffled.

Returns

img: The MultiChannelImage with shuffled pixel values.

Notes

This function iterates over the channels of the image, shuffles the pixel values in each channel using the shuffle! function from the Julia standard library, and stores the result back in the image.

The second method shuffles blocks of pixels in each channel of a MultiChannelImage. The block shuffling should retain the autocorrelation between neighboring pixels.

ProteinCoLoc.shuffle_blocks — Function

shuffle_blocks(img::MultiChannelImage, block_size::Int64)

This function shuffles the blocks of pixel values in each channel of a MultiChannelImage.

Arguments

img: A MultiChannelImage whose pixel blocks are to be shuffled.

block_size: An Int64 representing the size of the blocks to be shuffled.

Returns

img: The MultiChannelImage with shuffled pixel blocks.

Notes

This function calculates the number of blocks in each channel, retrieves the pixel values, shuffles the blocks, and stores the result back in the image. The shuffling is done by creating a new array with the first two dimensions shuffled. The blocks are then unpatched and the image data is updated.